Make group_train_test_split public and put it in a separate module
